+++ workspace/solutions/kotlin/lib/src/main/kotlin/_0/_0/_0/_1/s_0001_01/s_0001_01.kt
@@ -22,5 +22,3 @@ class Solution {
     }
 }
 // @lc code=end
-
-public class s_0001_01 : Solution