+++ workspace/solutions/kotlin/lib/src/test/kotlin/_0/_0/_0/_1/s_0001_test.kt
@@ -1,0 +1,16 @@
+package _0._0._0._1
+
+import utils.Test_
+import kotlin.test.Test
+import _0._0._0._1.s_0001_01.Solution as s_0001_01
+
+class s_0001_test {
+    @Test
+    fun test() {
+        val test = Test_(
+            this::class,
+            s_0001_01::class,
+        )
+        test.run()
+    }
+}